Refactoring and javadoc
- Move ILPSolver package
- Refactor code
- Add javadoc
- Use IbexOptions to define which ILPSolver to use
