--- org.emoflon.ibex.tgg.core.runtime/src/org/emoflon/ibex/common/ilp/gurobiWrapper/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author RobinO
- *
- */
-package org.emoflon.ibex.common.ilp.gurobiWrapper;
