+++ org.emoflon.ibex.tgg.core.runtime/src/org/emoflon/ibex/tgg/operational/strategies/OPT.java
@@ -17,13 +17,6 @@ import org.eclipse.emf.ecore.EPackage;
 import org.eclipse.emf.ecore.EReference;
 import org.eclipse.emf.ecore.impl.EClassImpl;
 import org.eclipse.emf.ecore.impl.EStructuralFeatureImpl;
-import org.emoflon.ibex.common.ilp.ILPFactory;
-import org.emoflon.ibex.common.ilp.ILPSolver;
-import org.emoflon.ibex.common.ilp.ILPFactory.SupportedILPSolver;
-import org.emoflon.ibex.common.ilp.ILPSolver.ILPLinearExpression;
-import org.emoflon.ibex.common.ilp.ILPSolver.ILPSolution;
-import org.emoflon.ibex.common.ilp.ILPSolver.ILPTerm;
-import org.emoflon.ibex.common.ilp.ILPSolver.Operation;
 import org.emoflon.ibex.tgg.compiler.patterns.sync.ConsistencyPattern;
 import org.emoflon.ibex.tgg.operational.defaults.IbexGreenInterpreter;
 import org.emoflon.ibex.tgg.operational.defaults.IbexOptions;
@@ -34,6 +27,13 @@ import org.emoflon.ibex.tgg.operational.strategies.cc.Bundle;
 import org.emoflon.ibex.tgg.operational.strategies.cc.ConsistencyReporter;
 import org.emoflon.ibex.tgg.operational.strategies.cc.HandleDependencies;
 import org.emoflon.ibex.tgg.operational.updatepolicy.IUpdatePolicy;
+import org.emoflon.ibex.tgg.util.ilp.ILPFactory;
+import org.emoflon.ibex.tgg.util.ilp.ILPSolver;
+import org.emoflon.ibex.tgg.util.ilp.ILPFactory.SupportedILPSolver;
+import org.emoflon.ibex.tgg.util.ilp.ILPSolver.ILPLinearExpression;
+import org.emoflon.ibex.tgg.util.ilp.ILPSolver.ILPSolution;
+import org.emoflon.ibex.tgg.util.ilp.ILPSolver.ILPTerm;
+import org.emoflon.ibex.tgg.util.ilp.ILPSolver.Operation;
 
 import com.google.common.collect.Sets;
 
@@ -72,59 +72,63 @@ public abstract class OPT extends OperationalStrategy {
 	 * value: kernels whose complement rules did not fulfill maximality property
 	 */
 	protected TIntHashSet invalidKernels = new TIntHashSet();
-	
+
 	/**
 	 * Collection of constraints to guarantee cyclic dependences are avoided;
 	 * value: correctly applied bundles (kernel match + its CRs matches)
 	 */
 	protected HashSet<Bundle> appliedBundles = new HashSet<Bundle>();
 	protected Bundle lastAppliedBundle;
-	
+
 	protected TIntObjectHashMap<String> matchIdToRuleName = new TIntObjectHashMap<>();
 	protected int idCounter = 1;
-	
+
 	//Hash maps to save the old metamodel state
 	THashMap<EReference, Integer> referenceToUpperBound = new THashMap<EReference, Integer>();
 	THashMap<EReference, Integer> referenceToLowerBound = new THashMap<EReference, Integer>();
 	THashMap<EReference, EReference> referenceToEOpposite = new THashMap<EReference, EReference>();
 	THashMap<EReference, Boolean> referenceToContainment = new THashMap<EReference, Boolean>();
 	
+	private final SupportedILPSolver ilpSolver;
+
 	public OPT(IbexOptions options) throws IOException {
 		super(options);
+		this.ilpSolver = options.getIlpSolver();
 	}
-	
+
 	public OPT(IbexOptions options, IUpdatePolicy policy) {
 		super(options, policy);
+		this.ilpSolver = options.getIlpSolver();
 	}
-	
+
 	public void relaxReferences(EList<EPackage> model) {
 
 		EPackage[] packages = (EPackage[])model.toArray();
-		
+
 		for (EPackage p : packages) {
 			TreeIterator<EObject> it = p.eAllContents();
-			
+
 			while (it.hasNext()) {
 				EObject next = it.next();
 				if (next instanceof EClassImpl) {
 					EClassImpl nextEClassImpl = (EClassImpl)next;
-					
+
 					for (EReference reference : nextEClassImpl.getEAllReferences()) {
 						if (referenceToUpperBound.containsKey(reference) &&
-							referenceToLowerBound.containsKey(reference) &&
-							referenceToContainment.containsKey(reference) &&
-							referenceToEOpposite.containsKey(reference)) {
+								referenceToLowerBound.containsKey(reference) &&
+								referenceToContainment.containsKey(reference) &&
+								referenceToEOpposite.containsKey(reference)) {
 							// Reference already exists, values must not be overwritten
-								continue;
+							continue;
 						}
-						
+
 						//Save metamodel values
 						referenceToUpperBound.put(reference, reference.getUpperBound());
 						referenceToLowerBound.put(reference, reference.getLowerBound());
 						referenceToContainment.put(reference, reference.isContainment());
 						referenceToEOpposite.put(reference, reference.getEOpposite());
-						
-						
+
+
 						//Change metamodel values
 						reference.setUpperBound(-1);
 						reference.setLowerBound(0);
@@ -135,32 +139,32 @@ public abstract class OPT extends OperationalStrategy {
 			}
 		}		
 	}
-	
+
 	public void unrelaxReferences(EList<EPackage> model) {
 
 		EPackage[] packages = (EPackage[])model.toArray();
-		
+
 		for (EPackage p : packages) {
 			TreeIterator<EObject> it = p.eAllContents();
-			
+
 			while (it.hasNext()) {
 				EObject next = it.next();
 				if (next instanceof EClassImpl) {
 					EClassImpl nextEClassImpl = (EClassImpl)next;
-					
+
 					for (EReference reference : nextEClassImpl.getEAllReferences()) {				
 						// Get old metamodel values
 						int upperBound = referenceToUpperBound.get(reference);
 						int lowerBound = referenceToLowerBound.get(reference);
 						boolean containment = referenceToContainment.get(reference);
 						EReference eOpposite = referenceToEOpposite.get(reference);
-						
+
 						// Change metamodel values
 						reference.setUpperBound(upperBound);
 						reference.setLowerBound(lowerBound);
 						reference.setContainment(containment);
 						reference.setEOpposite(eOpposite);
-						
+
 						// Reset setting for reference
 						((EStructuralFeatureImpl) reference).setSettingDelegate(null);
 					}
@@ -168,7 +172,7 @@ public abstract class OPT extends OperationalStrategy {
 			}
 		}
 	}
-	
+
 	protected void defineILPExclusions(ILPSolver ilpSolver) {
 		for (EObject node : nodeToMarkingMatches.keySet()) {
 			TIntHashSet variables = nodeToMarkingMatches.get(node);
@@ -191,7 +195,7 @@ public abstract class OPT extends OperationalStrategy {
 			ILPLinearExpression expr = ilpSolver.createLinearExpression(ilpTerms.toArray(new ILPTerm[0]));
 			ilpSolver.addConstraint(expr, Operation.le, 1.0, "EXCL" + nameCounter++);
 		}
-		
+
 		for (Integer match : sameCRmatches.keySet()) {
 			TIntHashSet variables = sameCRmatches.get(match);
 			List<ILPTerm> ilpTerms = new LinkedList<>();
@@ -202,7 +206,7 @@ public abstract class OPT extends OperationalStrategy {
 			ILPLinearExpression expr = ilpSolver.createLinearExpression(ilpTerms.toArray(new ILPTerm[0]));
 			ilpSolver.addConstraint(expr, Operation.le, 1.0, "EXCL" + nameCounter++);
 		}
-		
+
 		if (!invalidKernels.isEmpty()) {
 			TIntHashSet variables = invalidKernels;
 			List<ILPTerm> ilpTerms = new LinkedList<>();
@@ -213,7 +217,7 @@ public abstract class OPT extends OperationalStrategy {
 			ILPLinearExpression expr = ilpSolver.createLinearExpression(ilpTerms.toArray(new ILPTerm[0]));
 			ilpSolver.addConstraint(expr, Operation.le, 0.0, "EXCL" + nameCounter++);
 		}
-		
+
 		HandleDependencies handleCycles = new HandleDependencies(appliedBundles, edgeToMarkingMatches, nodeToMarkingMatches, matchToContextNodes, matchToContextEdges);
 		HashMap<Integer, ArrayList<Integer>> cyclicBundles = handleCycles.getCyclicDependenciesBetweenBundles();
 
@@ -280,36 +284,41 @@ public abstract class OPT extends OperationalStrategy {
 		ILPLinearExpression expr = ilpSolver.createLinearExpression(ilpTerms.toArray(new ILPTerm[0]));
 		ilpSolver.setObjective(expr, Operation.maximize);
 	}
-	
+
 	protected int[] chooseTGGRuleApplications() {
-		ILPSolver ilpSolver = ILPFactory.createBinaryILPSolver(SupportedILPSolver.Sat4J);		
+		ILPSolver ilpSolver = ILPFactory.createBinaryILPSolver(this.ilpSolver);		
 		defineILPExclusions(ilpSolver);
 		defineILPImplications(ilpSolver);
 		defineILPObjective(ilpSolver);
+		
+		try {
+			ILPSolution ilpSolution = ilpSolver.solveILP();
 
-		ILPSolution ilpSolution = ilpSolver.solveILP();
+			int[] result = new int[idToMatch.size()];
+			idToMatch.keySet().forEach(v -> {
+				if (ilpSolution.getVariable("x"+ v) > 0)
+					result[v - 1] = v;
+				else
+					result[v - 1] = -v;
+				return true;
+			});
 
-		int[] result = new int[idToMatch.size()];
-		idToMatch.keySet().forEach(v -> {
-			if (ilpSolution.getVariable("x"+ v) > 0)
-				result[v - 1] = v;
-			else
-				result[v - 1] = -v;
-			return true;
-		});
-
-		return result;
+			return result;
+		} catch (Exception e) {
+			e.printStackTrace();
+			return null;
+		}
 	}
-	
+
 	protected void handleBundles(IMatch comatch, String ruleName) {
 		if(!getComplementRule(ruleName).isPresent()) {
 			Bundle appliedBundle = new Bundle(idCounter);
 			appliedBundles.add(appliedBundle);
 			lastAppliedBundle = appliedBundle;
 		}
-		
+
 		lastAppliedBundle.addMatch(idCounter);
-		
+
 		// add context nodes and edges of this concrete match to its bundle
 		lastAppliedBundle.addBundleContextNodes(getBlackNodes(comatch, ruleName));
 		lastAppliedBundle.addBundleContextEdges(getBlackEdges(comatch, ruleName));
@@ -353,7 +362,7 @@ public abstract class OPT extends OperationalStrategy {
 		result.addAll(((IbexGreenInterpreter)greenInterpreter).createEdges(comatch, getGreenFactory(ruleName).getBlackTrgEdgesInRule(), false));
 		return result;
 	}
-	
+
 	protected Collection<EObject> getRuleApplicationNodes(IMatch comatch) {
 		return comatch.getParameterNames()
 				.stream()
