+++ org.emoflon.ibex.tgg.core.runtime/src/org/emoflon/ibex/tgg/util/ilp/package-info.java
@@ -1,0 +1,8 @@
+
+/**
+ * This package contains classes used for solving ILP problems by different ILPSolvers 
+ * 
+ * @author Robin Oppermann
+ *
+ */
+package org.emoflon.ibex.tgg.util.ilp;
